Move to shared hooks folder.
